handle: keep holder stale on detach
Simplify the handle->holder handling. From now on, we just rely on it
never being NULL except if never attached. This simplifies a lot of
places.

As a side effect, we no longer serialize on the holder. Instead, we use
n_weak, like we already do for remote-handles, anyway.

While at it, document the behavior and memory barriers.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>