peer: fix anchor destination lookup
If a fresh new anchor is used as destination, there are no weak
references, yet. Hence, we cannot look it up via the normal way. Do what
all the paths in handle.c already do: check whether it is an anchor, and
in this case just use the operating peer.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>